Convert float to double

+++ form_res_rels.c
@@ -32,7 +32,7 @@ static int comp_rank_judged (), comp_sim_docno (), comp_docno ();
 /* Definitions used for temporary and cached values */
 typedef struct {
     char *docno;
-    float sim;
+    double sim;
     long rank;
     long rel;
 } DOCNO_INFO;
